+++ projects/design-angular-kit/schematics/ng-add/rules/setup-project/index.ts
@@ -1,0 +1,6 @@
+export * from './add-animations';
+export * from './add-assets';
+export * from './add-design-angular-kit';
+export * from './add-http-client';
+export * from './add-import-to-style-file';
+export * from './add-localisation';