+++ projects/design-angular-kit/src/lib/components/core/tab/tab-container/tab-container.component.ts
@@ -4,8 +4,10 @@ import {
   Component,
   ContentChildren,
   ElementRef,
+  EventEmitter,
   Input,
   OnDestroy,
+  Output,
   QueryList,
   ViewChildren,
 } from '@angular/core';
@@ -42,13 +44,43 @@ export class ItTabContainerComponent extends ItAbstractComponent implements OnDe
   @Input({ transform: inputToBoolean }) dark?: boolean;
 
   /**
+   * Show items as cards
+   */
+  @Input({ transform: inputToBoolean }) cards?: boolean;
+
+  /**
+   * Show vertical navigation
+   */
+  @Input({ transform: inputToBoolean }) vertical?: boolean;
+
+  /**
+   * The tab position
+   */
+  @Input({ transform: inputToBoolean }) inverted?: boolean;
+
+  /**
+   * If tabs are editable
+   */
+  @Input({ transform: inputToBoolean }) editable?: boolean;
+
+  /**
    * The tab items
    */
   @ContentChildren(ItTabItemComponent) tabs?: QueryList<ItTabItemComponent>;
 
   @ViewChildren('tabNavLinks') private tabNavLinks?: QueryList<ElementRef<HTMLAnchorElement>>;
 
+  @Output() tabSelected = new EventEmitter<ItTabItemComponent>();
+
+  @Output() tabClosed = new EventEmitter();
+
+  @Output() tabAdded = new EventEmitter();
+
   private tabSubscriptions?: Array<Subscription>;
+
+  constructor() {
+    super();
+  }
 
   override ngAfterViewInit(): void {
     super.ngAfterViewInit();
@@ -88,5 +120,18 @@ export class ItTabContainerComponent extends ItAbstractComponent implements OnDe
 
   ngOnDestroy(): void {
     this.tabSubscriptions?.forEach(sub => sub.unsubscribe());
+  }
+
+  onTab(tab: ItTabItemComponent) {
+    this.tabSelected.emit(tab);
+  }
+
+  clickToClose(index: number) {
+    this.tabClosed.emit(index);
+  }
+
+  clickToAdd($event: Event) {
+    $event.preventDefault();
+    this.tabAdded.emit();
   }
 }